% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqUtility.R
\docType{methods}
\name{calcNormReadCountCoverage}
\alias{calcNormReadCountCoverage}
\title{Calculate Normalized Read Count Coverage}
\usage{
calcNormReadCountCoverage(coverageL, deseqds)
}
\arguments{
\item{coverageL}{output of calcReadCountCoveragePerCondition/generateCoverageDF_contigs}

\item{deseqds}{DESeq object of the current data}
}
\value{
data.frame containing the normalized read counts for each position in order to create a histogram of read counts for each ncRNA
}
\description{
This method returns a data frame containing normalized coverage values for each position. It requires the output of calcReadCountCoveragePerCondition or generateCoverageDF_contigs as data.frame or list
}

