% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{createSizeSelectedEnsemblncRNADB}
\alias{createSizeSelectedEnsemblncRNADB}
\title{Generate ncRNA database for rnastarMapping}
\usage{
createSizeSelectedEnsemblncRNADB(genomeDBdir = "/tmp/mmu", ncRNAsize = 400,
  nrThreads = 6, ensemblFTP = "mouse", generateIdx = FALSE)
}
\arguments{
\item{genomeDBdir}{Location to where the database should be saved}

\item{ncRNAsize}{maximum sequence length of the fasta file (default 400nt)}

\item{nrThreads}{Number of threads for the rnastar program (default 6)}

\item{ensemblFTP}{ftp adress e.g. ftp://ftp.ensembl.org/pub/release-78/fasta/mus_musculus/ncrna/, however for convenience for mouse and human the ftp sites are stored and
either 'mouse' or 'human' may be used for parametrization}

\item{generateIdx}{Specification if an rnastar index should be created (default FALSE)}
}
\value{
all commands generated
}
\description{
This method downloads ensembl ncRNA data if provided with ftp adress, modifies the fasta file,
subsets the fasta file to sequences of a given and generates a STAR index file if wanted.
}

