% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{extractRandomReadSet}
\alias{extractRandomReadSet}
\title{Randomly extract reads from fastq}
\usage{
extractRandomReadSet(numberOfReads = 50000, rawDataOldDir, rawDataNewDir)
}
\arguments{
\item{numberOfReads}{The number of reads one wants to extract (default: 50000)}

\item{rawDataOldDir}{The path to the fastq files from which the random subset is extracted}

\item{rawDataNewDir}{The path to the new sampled fastq files}
}
\value{
returns TRUE
}
\description{
Extract a random subset of reads from fastq files. The files have to be in the
same directory, since the method gets all fastq files from a directory and iterates through them.
}

