% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqUtility.R
\docType{methods}
\name{fetchNcRNAMappingSeq}
\alias{fetchNcRNAMappingSeq}
\title{Fetch sequences from ncRNAmapping}
\usage{
fetchNcRNAMappingSeq(ncRNAMappingDF, genomeIndexFilePath_ncRNA)
}
\arguments{
\item{ncRNAMappingDF}{candidates of interest}

\item{genomeIndexFilePath_ncRNA}{path to the fasta file where the STAR index is based on}
}
\value{
data.frame containing the UID and sequence
}
\description{
This method returns a data frame containing the sequence and UID from mapped ncRNAs
}

