% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{generateCountFromMappingDF}
\alias{generateCountFromMappingDF}
\title{Generate Count table}
\usage{
generateCountFromMappingDF(bamToBedAndMergencRNAL)
}
\arguments{
\item{bamToBedAndMergencRNAL}{bed files after mapping}
}
\value{
read count table
}
\description{
This method combines the read counts after counting them  with mergeBed from bam files
}

