% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqUtility.R
\docType{methods}
\name{generateCountUnclusteredTable}
\alias{generateCountUnclusteredTable}
\title{Helper method for clustering in MappAssClust_ncRNA_Template.R object}
\usage{
generateCountUnclusteredTable(contigFile, sampleBedFiles)
}
\arguments{
\item{contigFile}{CmdResult object is needed as input}

\item{sampleBedFiles}{CmdResult object is needed as input}
}
\value{
data.frame (Unclustered)
}
\description{
Not intended for general use, just a helper method for MappAssClust_ncRNA_Template.R
The method will merge the contig file generated from multiIntersect perl script with
all the contigs that have been intersected with -wb option to the multiIntersected file and merged afterwards.
Please see MappAssClust_ncRNA_Template.R, default inputs: multiIntersectBed_perl_CLI_cmdRes, mergeBedFile_CLI_cmdResL
}

