% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqUtility.R
\docType{methods}
\name{generateCoverageDF_contigs}
\alias{generateCoverageDF_contigs}
\title{Generate coverage data frame for specific candidates}
\usage{
generateCoverageDF_contigs(contigsGR, contigResTable, bedgraphMapping_plusL,
  bedgraphMapping_minusL, sampleNames, annotationGR)
}
\arguments{
\item{contigsGR}{GRanges file containing the coordinates of the contigs and their contigID (specified as contigID in the metadata)}

\item{contigResTable}{Selected candidates from the result of differential expression analysis from DESeq2 combined with the minimum annotation dataframe}

\item{bedgraphMapping_plusL}{the cmdResult object of the bedgraph plus mapping (genomeCovNcRNAMapping_plusL)}

\item{bedgraphMapping_minusL}{the cmdResult object of the bedgraph minus mapping (genomeCovNcRNAMapping_minusL)}

\item{sampleNames}{names of the samples from samplesInfo table}

\item{annotationGR}{An otpional annotation GRanges object, when provided the annotation will be used as reference interval}
}
\value{
data.frame containing the read counts for each position in order to create a histogram of read counts for each ncRNA
}
\description{
This method returns a data frame containing coverage values for each position. It is intended for genome mapping results. For ncRNA mappings please see: generateCoverageDF_ncRNAmappings
}

