% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqUtility.R
\docType{methods}
\name{splitReadCountCoveragePerCondition}
\alias{splitReadCountCoveragePerCondition}
\title{Splits coverage list per condition}
\usage{
splitReadCountCoveragePerCondition(coverageL, deseqds, samplesInfo)
}
\arguments{
\item{coverageL}{output of calcReadCountCoveragePerCondition/generateCoverageDF_contigs}

\item{deseqds}{DESeq object of the current data}

\item{samplesInfo}{describing the experiment (produced by configuration template)}
}
\value{
data.frame containing the normalized read counts for each position in order to create a histogram of read counts for each ncRNA
}
\description{
Splits the list obtained by calcReadCountCoveragePerCondition/generateCoverageDF_contigs by condition and calculates the mean read coverage per condition
}

