% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{summarizeContigAnnotation}
\alias{summarizeContigAnnotation}
\title{Summarize contig Annotation}
\usage{
summarizeContigAnnotation(contigDF)
}
\arguments{
\item{data.frame}{contig annotation table (contigAnnotTable_fin.csv)}
}
\value{
summarized contig annotation
}
\description{
This method summarizes the contig annotation table by first checking for range based annotation then for rfam annotation then for repeat annotation
}

