% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqUtility.R
\docType{methods}
\name{clusterMultiMappingReads_stringent_deprecated}
\alias{clusterMultiMappingReads_stringent_deprecated}
\title{Deprecated version of clustering}
\usage{
clusterMultiMappingReads_stringent_deprecated(contigForCountingGR_unclustered,
  allReads)
}
\arguments{
\item{contigForCountingGR_unclustered}{}

\item{allReads}{}
}
\value{
clusteredContigs
}
\description{
This method is deprecated do not use!
}

