% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqUtility.R
\docType{methods}
\name{generateCountFromMappingDF}
\alias{generateCountFromMappingDF}
\title{Generate Count table}
\usage{
generateCountFromMappingDF(bamToBedAndMergencRNAL, samplesInfo)
}
\arguments{
\item{bamToBedAndMergencRNAL}{bed files after mapping}
}
\value{
read count table
}
\description{
This method combines the read counts after counting them  with mergeBed from bam files
}

